+++ backend/GeneralMedicineOutlet.java
@@ -1,3 +1,5 @@
+package backend;
+
 import java.util.Date;
 import java.util.*;
 import java.text.SimpleDateFormat;
