+++ backend/MainDataConnection.java
@@ -1,3 +1,5 @@
+package backend;
+
 import java.sql.*;
 import java.io.*;
 import org.json.simple.*;
